Add few more charts
